requirements.txt
django

# Robots.txt
six                         #<-- Need By django-robots
django-robots

# Upload Images
django-filer

# Custom Wrap Errors
django-errors

# Needed by Heroku
django_heroku
gunicorn
